+++ CHANGELOG.md
@@ -69,6 +69,7 @@
 * nostr: add `Nip19Coordinate` struct ([Yuki Kishimoto])
 * nostr: add `Coordinate::verify` method ([Yuki Kishimoto])
 * nostr: add `TagStandard::Client` variant ([Yuki Kishimoto])
+* nostr: add `EventBuilder::dedup_tags` method ([Yuki Kishimoto])
 * pool: event verification cache ([Yuki Kishimoto])
 * ffi: add Mac Catalyst support in Swift package ([Yuki Kishimoto])